fix type of return in upsert method cause when property is created room create id with Long type

+++ app/src/main/java/com/openclassrooms/realestatemanager/database/Photo.kt
@@ -9,5 +9,5 @@ data class Photo (
     val id: Int,
     val shortDescription: String,
     val photoUri: String,
-    var propertyId: Int
+    var propertyId: Long
 )
